Several improvements/unification between the libCEED examples
- All examples except Nek support '-ceed' now as command-line parameter for the Ceed spec
- The MFEM and PETSc examples were renamed to bp?
- The libceed example (ex1-volume-annulus) was moved to examples/ceed/ex1
- For clarity, the libCEED portion of the MFEM examples were moved to header files
- Documentation + sample runs were added
