+++ examples/ceed/Makefile
@@ -15,7 +15,7 @@
 # testbed platforms, in support of the nation's exascale computing imperative.
 
 # Ceed directory
-CEED_DIR ?= ..
+CEED_DIR ?= ../..
 CEED_FLAGS ?= -I$(CEED_DIR)/include -std=c99
 CEED_LIBS ?= -Wl,-rpath,$(abspath $(CEED_DIR)/lib) -L$(CEED_DIR)/lib -lceed -lm
 
@@ -39,8 +39,8 @@ all: $(EXAMPLES)
 okl-lib-cache:
 # First test to see if libceed was compiled with OCCA
 ifneq ($(wildcard $(OCCA_DIR)/lib/libocca.*),)
-# Second test to see if '..' is CEED_DIR
-ifneq ($(CEED_DIR),..)
+# Second test to see if '../..' is CEED_DIR
+ifneq ($(CEED_DIR),../..)
 okl-lib-cache:
 	$(OCCA_DIR)/bin/occa cache ceed $(abspath $(CEED_DIR)/lib/okl/*.okl)
 endif
