+++ examples/petsc/bp1.c
@@ -1,6 +1,25 @@
-const char help[] = "Solve CEED BP1 using PETSc\n";
+//                        libCEED + PETSc Example: BP1
+//
+// This example demonstrates a simple usage of libCEED with PETSc to solve the
+// CEED BP1 benchmark problem, see http://ceed.exascaleproject.org/bps.
+//
+// The code is intentionally "raw", using only low-level communication
+// primitives.
+//
+// Build with:
+//
+//     make bp1 [PETSC_DIR=</path/to/mfem>] [CEED_DIR=</path/to/libceed>]
+//
+// Sample runs:
+//
+//     bp1
+//     bp1 -ceed /cpu/self
+//     bp1 -ceed /gpu/occa
+//     bp1 -ceed /cpu/occa
+//     bp1 -ceed /omp/occa
+//     bp1 -ceed /ocl/occa
 
-// This example is intentionally "raw", using only low-level communication primitives from PETSc
+const char help[] = "Solve CEED BP1 using PETSc\n";
 
 #include <petscksp.h>
 #include <ceed.h>
@@ -28,7 +47,7 @@ static void GlobalDof(const PetscInt p[3], const PetscInt irank[3],
     mdof[d] = degree*melem[d] + (irank[d] == p[d]-1);
 }
 static PetscInt GlobalStart(const PetscInt p[3], const PetscInt irank[3],
-                        PetscInt degree, const PetscInt melem[3]) {
+                            PetscInt degree, const PetscInt melem[3]) {
   PetscInt start = 0;
   // Dumb brute-force is easier to read
   for (PetscInt i=0; i<p[0]; i++) {
@@ -44,8 +63,8 @@ static PetscInt GlobalStart(const PetscInt p[3], const PetscInt irank[3],
   return -1;
 }
 static int CreateRestriction(Ceed ceed, const CeedInt melem[3],
-                                        CeedInt P,
-                                        CeedElemRestriction *Erestrict) {
+                             CeedInt P,
+                             CeedElemRestriction *Erestrict) {
   const PetscInt Nelem = melem[0]*melem[1]*melem[2];
   PetscInt mdof[3], *idx, *idxp;
 
@@ -78,18 +97,19 @@ static int CreateRestriction(Ceed ceed, const CeedInt melem[3],
 }
 
 static int Setup(void *ctx, void *qdata, CeedInt Q,
-                const CeedScalar *const *in, CeedScalar *const *out) {
+                 const CeedScalar *const *in, CeedScalar *const *out) {
   CeedScalar *rho = qdata, *target = rho + Q;
   const CeedScalar (*x)[Q] = (const CeedScalar (*)[Q])in[0],
-    (*J)[3][Q] = (const CeedScalar (*)[3][Q])in[1],
-    *w = in[4];
+                             (*J)[3][Q] = (const CeedScalar (*)[3][Q])in[1],
+                                          *w = in[4];
   for (CeedInt i=0; i<Q; i++) {
     CeedScalar det =
-        J[0][0][i] * (J[1][1][i]*J[2][2][i] - J[1][2][i]*J[2][1][i])
+      J[0][0][i] * (J[1][1][i]*J[2][2][i] - J[1][2][i]*J[2][1][i])
       - J[0][1][i] * (J[1][0][i]*J[2][2][i] - J[1][2][i]*J[2][0][i])
       + J[0][2][i] * (J[1][0][i]*J[2][1][i] - J[1][1][i]*J[2][0][i]);
     rho[i] = det * w[i];
-    target[i] = PetscSqrtScalar(PetscSqr(x[0][i]) + PetscSqr(x[1][i]) + PetscSqr(x[2][i]));
+    target[i] = PetscSqrtScalar(PetscSqr(x[0][i]) + PetscSqr(x[1][i]) + PetscSqr(
+                                  x[2][i]));
   }
   return 0;
 }
@@ -105,7 +125,7 @@ static int Mass(void *ctx, void *qdata, CeedInt Q,
 }
 
 static int Error(void *ctx, void *qdata, CeedInt Q,
-                const CeedScalar *const *in, CeedScalar *const *out) {
+                 const CeedScalar *const *in, CeedScalar *const *out) {
   CeedScalar *maxerror = ctx;
   CeedScalar *rho = qdata, *target = rho + Q;
   for (CeedInt i=0; i<Q; i++) {
@@ -131,13 +151,15 @@ static PetscErrorCode MatMult_Mass(Mat A, Vec X, Vec Y) {
   PetscScalar *x, *y;
 
   PetscFunctionBeginUser;
-  ierr = MatShellGetContext(A, &user);CHKERRQ(ierr);
-  ierr = VecScatterBegin(user->ltog, X, user->Xloc, INSERT_VALUES, SCATTER_REVERSE);CHKERRQ(ierr);
-  ierr = VecScatterEnd(user->ltog, X, user->Xloc, INSERT_VALUES, SCATTER_REVERSE);CHKERRQ(ierr);
-  ierr = VecZeroEntries(user->Yloc);CHKERRQ(ierr);
+  ierr = MatShellGetContext(A, &user); CHKERRQ(ierr);
+  ierr = VecScatterBegin(user->ltog, X, user->Xloc, INSERT_VALUES,
+                         SCATTER_REVERSE); CHKERRQ(ierr);
+  ierr = VecScatterEnd(user->ltog, X, user->Xloc, INSERT_VALUES, SCATTER_REVERSE);
+  CHKERRQ(ierr);
+  ierr = VecZeroEntries(user->Yloc); CHKERRQ(ierr);
 
-  ierr = VecGetArrayRead(user->Xloc, (const PetscScalar**)&x);CHKERRQ(ierr);
-  ierr = VecGetArray(user->Yloc, &y);CHKERRQ(ierr);
+  ierr = VecGetArrayRead(user->Xloc, (const PetscScalar**)&x); CHKERRQ(ierr);
+  ierr = VecGetArray(user->Yloc, &y); CHKERRQ(ierr);
   CeedVectorSetArray(user->xceed, CEED_MEM_HOST, CEED_USE_POINTER, x);
   CeedVectorSetArray(user->yceed, CEED_MEM_HOST, CEED_USE_POINTER, y);
 
@@ -149,8 +171,10 @@ static PetscErrorCode MatMult_Mass(Mat A, Vec X, Vec Y) {
 
   if (Y) {
     ierr = VecZeroEntries(Y); CHKERRQ(ierr);
-    ierr = VecScatterBegin(user->ltog, user->Yloc, Y, ADD_VALUES, SCATTER_FORWARD); CHKERRQ(ierr);
-    ierr = VecScatterEnd(user->ltog, user->Yloc, Y, ADD_VALUES, SCATTER_FORWARD); CHKERRQ(ierr);
+    ierr = VecScatterBegin(user->ltog, user->Yloc, Y, ADD_VALUES, SCATTER_FORWARD);
+    CHKERRQ(ierr);
+    ierr = VecScatterEnd(user->ltog, user->Yloc, Y, ADD_VALUES, SCATTER_FORWARD);
+    CHKERRQ(ierr);
   }
   PetscFunctionReturn(0);
 }
@@ -160,7 +184,7 @@ int main(int argc, char **argv) {
   MPI_Comm comm;
   char ceedresource[4096] = "/cpu/self";
   PetscInt degree, qextra, localdof, localelem, melem[3], mdof[3], p[3],
-    irank[3], ldof[3], lsize;
+           irank[3], ldof[3], lsize;
   PetscMPIInt size, rank;
   VecScatter ltog;
   Ceed ceed;
@@ -175,7 +199,8 @@ int main(int argc, char **argv) {
   KSP ksp;
   User user;
 
-  ierr = PetscInitialize(&argc, &argv, NULL, help); if (ierr) return ierr;
+  ierr = PetscInitialize(&argc, &argv, NULL, help);
+  if (ierr) return ierr;
   comm = PETSC_COMM_WORLD;
   ierr = PetscOptionsBegin(comm, NULL, "CEED BP1 in PETSc", NULL); CHKERRQ(ierr);
   degree = 1;
@@ -184,20 +209,22 @@ int main(int argc, char **argv) {
   qextra = 0;
   ierr = PetscOptionsInt("-qextra", "Number of extra quadrature points",
                          NULL, qextra, &qextra, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsString("-ceed_resource", "CEED resource specifier",
+  ierr = PetscOptionsString("-ceed", "CEED resource specifier",
                             NULL, ceedresource, ceedresource,
                             sizeof(ceedresource), NULL); CHKERRQ(ierr);
   localdof = 1000;
-  ierr = PetscOptionsInt("-local", "Target number of locally owned degrees of freedom per process",
+  ierr = PetscOptionsInt("-local",
+                         "Target number of locally owned degrees of freedom per process",
                          NULL, localdof, &localdof, NULL); CHKERRQ(ierr);
-  ierr = PetscOptionsEnd();CHKERRQ(ierr);
+  ierr = PetscOptionsEnd(); CHKERRQ(ierr);
 
   // Determine size of process grid
   ierr = MPI_Comm_size(comm, &size); CHKERRQ(ierr);
   Split3(size, p, false);
 
   // Find a nicely composite number of elements no less than localdof
-  for (localelem = PetscMax(1, localdof / (degree*degree*degree)); ; localelem++) {
+  for (localelem = PetscMax(1, localdof / (degree*degree*degree)); ;
+       localelem++) {
     Split3(localelem, melem, true);
     if (Max3(melem) / Min3(melem) <= 2) break;
   }
@@ -221,7 +248,7 @@ int main(int argc, char **argv) {
 
   {
     ierr = VecCreate(comm, &X); CHKERRQ(ierr);
-    ierr = VecSetSizes(X, mdof[0] * mdof[1] * mdof[2], PETSC_DECIDE);CHKERRQ(ierr);
+    ierr = VecSetSizes(X, mdof[0] * mdof[1] * mdof[2], PETSC_DECIDE); CHKERRQ(ierr);
     ierr = VecSetUp(X); CHKERRQ(ierr);
 
     lsize = 1;
@@ -230,7 +257,7 @@ int main(int argc, char **argv) {
       lsize *= ldof[d];
     }
     ierr = VecCreate(PETSC_COMM_SELF, &Xloc); CHKERRQ(ierr);
-    ierr = VecSetSizes(Xloc, lsize, PETSC_DECIDE);CHKERRQ(ierr);
+    ierr = VecSetSizes(Xloc, lsize, PETSC_DECIDE); CHKERRQ(ierr);
     ierr = VecSetUp(Xloc); CHKERRQ(ierr);
 
     // Create local-to-global scatter
@@ -253,12 +280,14 @@ int main(int argc, char **argv) {
       for (PetscInt j=0,jr,jj; jr=j>=mdof[1], jj=j-jr*mdof[1], j<ldof[1]; j++) {
         for (PetscInt k=0,kr,kk; kr=k>=mdof[2], kk=k-kr*mdof[2], k<ldof[2]; k++) {
           ltogind[(i*ldof[1]+j)*ldof[2]+k] = gstart[ir][jr][kr]
-            + (ii*gmdof[ir][jr][kr][1]+jj)*gmdof[ir][jr][kr][2]+kk;
+                                             + (ii*gmdof[ir][jr][kr][1]+jj)*gmdof[ir][jr][kr][2]+kk;
         }
       }
     }
-    ierr = ISCreateGeneral(comm, lsize, ltogind, PETSC_OWN_POINTER, &ltogis); CHKERRQ(ierr);
-    ierr = VecScatterCreate(Xloc, NULL, X, ltogis, &ltog); CHKERRQ(ierr); CHKERRQ(ierr);
+    ierr = ISCreateGeneral(comm, lsize, ltogind, PETSC_OWN_POINTER, &ltogis);
+    CHKERRQ(ierr);
+    ierr = VecScatterCreate(Xloc, NULL, X, ltogis, &ltog); CHKERRQ(ierr);
+    CHKERRQ(ierr);
     ierr = ISDestroy(&ltogis); CHKERRQ(ierr);
   }
 
@@ -272,14 +301,18 @@ int main(int argc, char **argv) {
   CreateRestriction(ceed, melem, 2, &Erestrictx);
   {
     CeedScalar *xloc;
-    CeedInt shape[3] = {melem[0]+1, melem[1]+1, melem[2]+1}, len = shape[0]*shape[1]*shape[2];
+    CeedInt shape[3] = {melem[0]+1, melem[1]+1, melem[2]+1}, len =
+                         shape[0]*shape[1]*shape[2];
     xloc = malloc(len*3*sizeof xloc[0]);
     for (CeedInt i=0; i<shape[0]; i++) {
       for (CeedInt j=0; j<shape[1]; j++) {
         for (CeedInt k=0; k<shape[2]; k++) {
-          xloc[((i*shape[1]+j)*shape[2]+k) + 0*len] = 1.*(irank[0]*melem[0]+i) / (p[0]*melem[0]);
-          xloc[((i*shape[1]+j)*shape[2]+k) + 1*len] = 1.*(irank[1]*melem[1]+j) / (p[1]*melem[1]);
-          xloc[((i*shape[1]+j)*shape[2]+k) + 2*len] = 1.*(irank[2]*melem[2]+k) / (p[2]*melem[2]);
+          xloc[((i*shape[1]+j)*shape[2]+k) + 0*len] = 1.*(irank[0]*melem[0]+i) /
+              (p[0]*melem[0]);
+          xloc[((i*shape[1]+j)*shape[2]+k) + 1*len] = 1.*(irank[1]*melem[1]+j) /
+              (p[1]*melem[1]);
+          xloc[((i*shape[1]+j)*shape[2]+k) + 2*len] = 1.*(irank[2]*melem[2]+k) /
+              (p[2]*melem[2]);
         }
       }
     }
@@ -317,7 +350,8 @@ int main(int argc, char **argv) {
 
   ierr = MatCreateShell(comm, mdof[0]*mdof[1]*mdof[2], mdof[0]*mdof[1]*mdof[2],
                         PETSC_DECIDE, PETSC_DECIDE, user, &mat); CHKERRQ(ierr);
-  ierr = MatShellSetOperation(mat, MATOP_MULT, (void(*)(void))MatMult_Mass); CHKERRQ(ierr);
+  ierr = MatShellSetOperation(mat, MATOP_MULT, (void(*)(void))MatMult_Mass);
+  CHKERRQ(ierr);
 
   ierr = MatCreateVecs(mat, &rhs, NULL); CHKERRQ(ierr);
   ierr = VecZeroEntries(X); CHKERRQ(ierr);
@@ -335,7 +369,8 @@ int main(int argc, char **argv) {
     ierr = PCSetType(pc, PCJACOBI); CHKERRQ(ierr);
     ierr = PCJacobiSetType(pc, PC_JACOBI_ROWSUM); CHKERRQ(ierr);
     ierr = KSPSetType(ksp, KSPCG); CHKERRQ(ierr);
-    ierr = KSPSetTolerances(ksp, 1e-10, PETSC_DEFAULT, PETSC_DEFAULT, PETSC_DEFAULT); CHKERRQ(ierr);
+    ierr = KSPSetTolerances(ksp, 1e-10, PETSC_DEFAULT, PETSC_DEFAULT,
+                            PETSC_DEFAULT); CHKERRQ(ierr);
   }
   ierr = KSPSetFromOptions(ksp); CHKERRQ(ierr);
   ierr = KSPSetOperators(ksp, mat, mat); CHKERRQ(ierr);
@@ -357,7 +392,8 @@ int main(int argc, char **argv) {
   user->op = op_error;
   CeedQFunctionSetContext(qf_error, &maxerror, sizeof maxerror);
   ierr = MatMult_Mass(mat, X, NULL); CHKERRQ(ierr);
-  ierr = PetscPrintf(comm, "Pointwise error (max) %e\n", (double)maxerror); CHKERRQ(ierr);
+  ierr = PetscPrintf(comm, "Pointwise error (max) %e\n", (double)maxerror);
+  CHKERRQ(ierr);
 
   ierr = VecDestroy(&rhs); CHKERRQ(ierr);
   ierr = VecDestroy(&X); CHKERRQ(ierr);
